marking_assistant/requirements.txt
pydantic>=2.10
pydantic-ai>=0.2.17
pandas>=1.5.0
openpyxl>=3.0.0
pyyaml>=6.0
pathlib
typing-extensions
argparse 